backend/requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
Jinja2==3.1.4
MarkupSafe==2.1.5
itsdangerous==2.2.0
click==8.1.7
blinker==1.8.2
python-docx==1.1.2
gunicorn==21.2.0
lxml==5.3.0
openpyxl==3.1.2
